Move Environment class to its own module
Making `tmt.utils` smaller, but also part of the secrets implementation:
by moving `Environment` and few other bits, it should be possible to
close access to `os.environ`.

A lot of changes, but just moves - some `tmt.utils.Environment` turned
into `from tmt.utils.environment import Environment` + bare
`Environment`; the string would have to change anyway, so at least we
can treat `Environment` as a well-known name in the codebase.

Related to #2609.
