+++ tmt/checks/internal/timeout.py
@@ -1,13 +1,13 @@
 from typing import Optional
 
 import tmt.log
-import tmt.utils
 from tmt.checks import CheckPlugin, provides_check
 from tmt.checks.internal import InternalCheck
 from tmt.container import container
 from tmt.result import CheckResult, ResultOutcome
 from tmt.steps.execute import TestInvocation
 from tmt.utils import ProcessExitCodes
+from tmt.utils.environment import Environment
 
 CHECK_NAME = 'internal/timeout'
 
@@ -40,7 +40,7 @@ class Timeout(CheckPlugin[TimeoutCheck]):
         *,
         check: 'TimeoutCheck',
         invocation: 'TestInvocation',
-        environment: Optional[tmt.utils.Environment] = None,
+        environment: Optional[Environment] = None,
         logger: tmt.log.Logger,
     ) -> list[CheckResult]:
         if invocation.return_code == ProcessExitCodes.TIMEOUT:
