+++ tmt/steps/execute/upgrade.py
@@ -21,7 +21,8 @@ from tmt.steps.execute import ExecutePlugin
 from tmt.steps.execute.internal import ExecuteInternal, ExecuteInternalData
 from tmt.steps.prepare import PreparePlugin
 from tmt.steps.prepare.install import PrepareInstallData
-from tmt.utils import Environment, EnvVarValue, Path
+from tmt.utils import Path
+from tmt.utils.environment import Environment, EnvVarValue
 
 STATUS_VARIABLE = 'IN_PLACE_UPGRADE'
 BEFORE_UPGRADE_PREFIX = 'old'
@@ -281,7 +282,7 @@ class ExecuteUpgrade(ExecuteInternal):
         self,
         *,
         guest: tmt.guest.Guest,
-        environment: Optional[tmt.utils.Environment] = None,
+        environment: Optional[Environment] = None,
         logger: tmt.log.Logger,
     ) -> None:
         """
