+++ tmt/steps/finish/__init__.py
@@ -17,6 +17,9 @@ from tmt.steps import (
     PullTask,
     sync_with_guests,
 )
+from tmt.utils.environment import Environment
+from tmt.utils.environment import EnvVarName as EnvVarName
+from tmt.utils.environment import EnvVarValue as EnvVarValue
 
 
 @container
@@ -43,7 +46,7 @@ class FinishPlugin(tmt.steps.Plugin[FinishStepDataT, PluginOutcome]):
         self,
         *,
         guest: 'Guest',
-        environment: Optional[tmt.utils.Environment] = None,
+        environment: Optional[Environment] = None,
         logger: tmt.log.Logger,
     ) -> PluginOutcome:
         self.go_prolog(logger)
