+++ tmt/steps/prepare/artifact/__init__.py
@@ -25,7 +25,10 @@ from tmt.steps.prepare.verify_installation import (
     PrepareVerifyInstallation,  # pyright: ignore[reportUnknownVariableType]
     PrepareVerifyInstallationData,
 )
-from tmt.utils import Environment, Path
+from tmt.utils import Path
+from tmt.utils.environment import Environment
+from tmt.utils.environment import EnvVarName as EnvVarName
+from tmt.utils.environment import EnvVarValue as EnvVarValue
 
 
 @container
